fix: read notes
Home and Stats page "Total stats" are not live anymore. This was actually incredibly heavy on my database, updating every 5 seconds times whatever number of users were online, it was getting pretty bad.
So this is now updated every 5 minutes on the database, and for the website is static.
I really liked how it worked though and I plan to add some kind of "fake live" total stats updating at a later time where the average totals are used to make a fake live counter that is pretty close to what it should be